fullfillment-python/requirements.txt
Flask
requests
google-cloud-bigquery
